requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
python-dotenv
httpx
python-docx
PyPDF2
python-multipart
